Split some functions and tests into differnt files (#7684)
