+++ tfplan2cai/converters/google/resources/validation.go
@@ -355,3 +355,31 @@ func validateADDomainName() schema.SchemaValidateFunc {
 		return
 	}
 }
+
+func testStringValidationCases(cases []StringValidationTestCase, validationFunc schema.SchemaValidateFunc) []error {
+	es := make([]error, 0)
+	for _, c := range cases {
+		es = append(es, testStringValidation(c, validationFunc)...)
+	}
+
+	return es
+}
+
+func testStringValidation(testCase StringValidationTestCase, validationFunc schema.SchemaValidateFunc) []error {
+	_, es := validationFunc(testCase.Value, testCase.TestName)
+	if testCase.ExpectError {
+		if len(es) > 0 {
+			return nil
+		} else {
+			return []error{fmt.Errorf("Didn't see expected error in case \"%s\" with string \"%s\"", testCase.TestName, testCase.Value)}
+		}
+	}
+
+	return es
+}
+
+type StringValidationTestCase struct {
+	TestName    string
+	Value       string
+	ExpectError bool
+}
